[Kernel][Helion] Split generated dispatch by op
